update less task to work as expected
this is still not perfect but more close to what is expected.

In order to have local less files compiled against a default theme, you
would need to provide a `coreDir` variable. This can be provided via
command line or via a file in `grunt/local.conf.json`. This `coreDir`
variable needs to point to a directory containing core less files, like
a distributed version of the core UI or the core UI repository itself.