don't touch mode of files during install tasks
during installation, files should be copied from dist folder as is. This
allows the package to ship binary files with executable flag set.
Modules that want to make use of this, would need to adjust their dist
tasks, in order to have the correct flags in dist/ directory, so install
tasks can pick those up.