more strict copying for dist task
when copy:dist is run, also js files are copied. Since it is
theoretically possible, that uglify task is run before copy:dist, this
will (and has) lead to problems. Be more strict about which files should
be handled when exactly. This also brings an improvement, because all
files ending with .min.js will not be handled by uglify task, speeding
things up, if there are a lot of large, already minified libraries in
the build/ directory.