change behaviour of copying files from apps/ to build/
after some discussion, we decided to change the behaviour of copying
files from the apps/ directory to build/. This change simplifies the
configuration in some cases but leads to a higher possibility of
unwanted files being copied during copy_build task.

So the new behaviour is to copy **everything** from the apps/ directory
and not only some pre-defined file types. There is no longer the need to
extend the configuration, if you only want to copy files from the apps/
directory. If you used this feature before (like, you did some
pre-processing of the files in apps/ and wanted these files not to be
copied), we suggest to move such files outside the apps/ directory and
provide a completely custom task for such files.