fix problem with bower task
the bower task is supposed to install bower packages and not to let
grunt fail. However, grunt-check-dependencies exits grunt with an exit
code other than 0, if bower install has been invoked.

This works around this issue and restores old behaviour. There is a pull
request for upstream, that makes this behaviour configurable. See
https://github.com/mzgol/grunt-check-dependencies/pull/8 for details.