support custom https key configuration for connect
To get rid of those ssl warnings in the browser, one way is to supply
correct ssl certificates. It's now possible to maintain such
certificates (one might want to use wild-card certificates) in a team
and put those in the ssl/ directory of the project. If those files are
okay, everything will work as expected. Single developers can override
those values in grunt/local.conf.json.