add optional proxy server
this can be used to start a proxy server. This can be configured in the
browser, which will then send all requests through the proxy server. For
SSL connections, this will use the HTTP CONNECT method. For HTTP
traffic, requests will be forward proxied directly.

For HTTPS traffic, there is one exception: the server configured in the
appserver section will be treated in a special way. Those requests will
go through the appserver stack and so some of the requests might be
served completely locally and others might get forwarded to the actual
backend.

This can be used as a convenient way to test frontend code against
complex backend setups involving a lot of machines with different
hostnames, e.g. for login handling or other things.