fix i18n module template to handle plural forms correctly
languages like jp_JP that only have one plural form, so the test for the
number of translations is not correct. Correct would be to test for the
magic character \x01, which is used as a separator between the msgid and
msgid_plural. So \x01 is the real indicator, wether a string is in
plural form or not.